Revert "2nd enemy sprite"
