--- Assets/Character/Sprites/Enemies.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9be91114fed6b3947b2f6829412c4479
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
