--- Assets/Character/Sprites/Enemies/RedSlime.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 54b40203f178ce045a9c674b08eaff9a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
