--- Assets/Character/Sprites/Enemies/RedSlime/RedSlimeController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b0be3af26306f654093f0cfa052d8fba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
