--- Assets/Character/Sprites/Enemies/SniperJoe.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: af1907134891ee74e9b05e9cc62392e1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
