--- Assets/Character/Sprites/Enemies/SniperJoe/SniperJoeAnimationController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c7d89949ac4305e4ab36c888a2a7eed7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
