--- Assets/Character/Sprites/SniperJoe.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0cd0f4b6e6b076844ab19e1ba5bec24b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
