--- Assets/Character/Sprites/SniperJoe/Idle1.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c3e2ec43584fc6bf067461af9115a2a399e797dcfd057084783f98b4f6836a17
-size 1217
