--- Assets/Character/Sprites/SniperJoe/Idle2.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:eb635d9149291361e421c55c96a8c5239438fec28f42dff29717e73d38458591
-size 1372
