--- Assets/Character/Sprites/SniperJoe/PreShoot.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5cc505fff080e4fe030dc793ac78ce471bd97bfc709713238b776a0b0b593735
-size 1133
