--- Assets/Character/Sprites/SniperJoe/Shoot.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:649d32e89ebeb3dd46a73085a90fe372e16906d84571a5eb39aefa1609c97140
-size 1208
