--- Assets/Environment/Sprites/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 102f3ea9217104a4992869a47f4a8e20
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
