--- Assets/Environment/Tiles/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2b75c907b4b3b134b8d45bf2f3a7a52b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
