--- Assets/Scripts/Foreground_Canvas.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9c7a3a694a0b32c49b1f8f5fab4fe67e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
