+++ Assets/Scripts/Gameplay/EnemyDeath.cs
@@ -1,4 +1,3 @@
-using Assets.Scripts;
 using Platformer.Core;
 using Platformer.Mechanics;
 
@@ -10,11 +9,12 @@ namespace Platformer.Gameplay
     /// <typeparam name="EnemyDeath"></typeparam>
     public class EnemyDeath : Simulation.Event<EnemyDeath>
     {
-        public IEnemy enemy;
+        public EnemyController enemy;
 
         public override void Execute()
         {
-            enemy.SetCollision(false); 
+            enemy._collider.enabled = false;
+            enemy.control.enabled = false;
             enemy.health.Die();
             if (enemy._audio && enemy.ouch)
                 enemy._audio.PlayOneShot(enemy.ouch);
