--- Assets/Scripts/Mechanics/SniperJoeController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 832d70bf66840224cbfa2e30a84ff2d9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
