--- Assets/Weapons/EnemyBlaster.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6a4e3e6232d9a724b9aa2a7782961e83
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
