--- Assets/Weapons/EnemyBullet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6e2f6339afdba504193f55516e068876
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
