--- Assets/Weapons/EnemyBullet.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4367c03f780d79248bc681d2ca296ac7
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
