--- Assets/Weapons/PlayerBullet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e0499b29c170be447bd291e6fa6d3bbf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
