--- Assets/Weapons/PlayerWeapon.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d79cd4fde10e68f46a423610a2bd792b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
