--- Assets/Weapons/SniperJoeWeapon.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4bbda4e7a41b95f4ca570eda13c38811
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
