+++ Assets/Weapons/Weapon.cs
@@ -1,47 +1,97 @@
-﻿using Assets.Weapons;
-using Platformer.Mechanics;
+﻿using Platformer.Mechanics;
 using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using UnityEngine;
 
-public abstract class Weapon<T> :MonoBehaviour where T :Bullet
+public class Weapon : MonoBehaviour
 {
     public Transform firePoint;
     public GameObject bulletPrefabNormal;
+    public GameObject bulletPrefabCharged1;
+    public GameObject bulletPrefabCharged2;
+    public PlayerController player;
+    private bool charging = false;
+    private Stopwatch chargeStopwatch = new Stopwatch();
+    public ChargingState ChargeState;
 
     public float FireCooldownSeconds;
     public int MaxBullets;
-    protected float deltaTime;
-
-    public bool IsEnabled { get; set; }
+    private float deltaTime;
 
     private void Start()
     {
-        IsEnabled = true;
         deltaTime = FireCooldownSeconds; //let you shoot immediately
     }
 
     // Update is called once per frame
-    protected virtual void Update()
+    void Update()
     {
-        if(IsEnabled)
+        bool shooting = false;
+        deltaTime += Time.deltaTime;
+        if (Input.GetButtonDown("Fire1"))
         {
-            deltaTime += Time.deltaTime;
+            ChargeState = ChargingState.Tier0;
             Shoot(1);
+            shooting = true;
+        }
+        if (Input.GetButton("Fire1") && !charging)
+        {
+            charging = true;
+            chargeStopwatch.Restart();
+        }
+        if (Input.GetButtonUp("Fire1") && charging)
+        {
+            charging = false;
+            chargeStopwatch.Stop();
+            switch(ChargeState)
+            {
+                case ChargingState.Tier0:
+                    Shoot(1);
+                    break;
+                case ChargingState.Tier1:
+                    Shoot(2);
+                    break;
+                case ChargingState.Tier2:
+                    Shoot(3);
+                    break;
+            }
+            shooting = true;
+        }
+
+        player.animationController.SetShooting(shooting);
+    }
+
+    private void FixedUpdate()
+    {
+        if (charging)
+        {
+            long chargedMS = chargeStopwatch.ElapsedMilliseconds;
+            if (chargedMS > 2500)
+                ChargeState = ChargingState.Tier2;
+            else if (chargedMS > 1000)
+                ChargeState = ChargingState.Tier1;
+        }
+        else
+        {
+            ChargeState = ChargingState.Tier0;
         }
     }
 
-    protected void Shoot(float power)
+    private void Shoot(float power)
     {
         if (deltaTime < FireCooldownSeconds)
             return;
 
-        GameObject prefab = GetPrefabBullet(power);
+        GameObject prefab = bulletPrefabNormal;
+        if (power == 2)
+            prefab = bulletPrefabCharged1;
+        else if (power == 3)
+            prefab = bulletPrefabCharged2;
 
         GameObject obj = Instantiate(prefab, firePoint.position, firePoint.rotation);
-        T bullet = obj.GetComponent<T>();
+        Bullet bullet = obj.GetComponent<Bullet>();
         if (bullet != null)
         {
             bullet.SetPower(power);
@@ -49,9 +99,10 @@ public abstract class Weapon<T> :MonoBehaviour where T :Bullet
         deltaTime = power <= 1 ? 0 : FireCooldownSeconds;
     }
 
-    protected virtual GameObject GetPrefabBullet(float power)
+    public enum ChargingState
     {
-        GameObject prefab = bulletPrefabNormal;
-        return prefab;
+        Tier0,
+        Tier1,
+        Tier2
     }
 }
