+++ Packages/manifest.json
@@ -9,7 +9,7 @@
     "com.unity.ide.vscode": "1.2.0",
     "com.unity.learn.iet-framework": "0.3.0-preview.5",
     "com.unity.postprocessing": "2.3.0",
-    "com.unity.test-framework": "1.1.14",
+    "com.unity.test-framework": "1.1.13",
     "com.unity.textmeshpro": "2.0.1",
     "com.unity.timeline": "1.2.14",
     "com.unity.ugui": "1.0.0",
